[WIT-109] Helm chart for java scaffold
# New features and improvements

This MR introduces the helm chart for deploying the java scaffold.

# Bug fixes

**Maven  package**

Adds a maven goal to properly set the Main class in the jar.

**Container scanning**

The container-scanning job is scanning the "latest" docker image, not the one produced by the very same cicd. This MR specifies the exact image to scan for vulnerabilities.

# Related issue

Closes WIT-109